extract formSegment from summon-app; replace in all instances in summon
