+++ apps/summon-app/src/layouts/AdvancedSegment.tsx
@@ -1,6 +1,6 @@
-import { WrappedInput } from '@daohaus/ui';
+import { FormSegment, WrappedInput } from '@daohaus/ui';
 
-import { FormSegment, SplitColumn } from '../layouts/FormLayouts';
+import { SplitColumn } from '../layouts/FormLayouts';
 import { INFO_COPY } from '../utils/content';
 import { FORM_KEYS } from '../utils/formKeys';
