+++ apps/summon-app/src/layouts/MemberSegment.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from 'react';
 import { useFormContext } from 'react-hook-form';
 
-import { ParSm, WrappedTextArea } from '@daohaus/ui';
+import { FormSegment, ParSm, WrappedTextArea } from '@daohaus/ui';
 
-import { FormSegment, TextAreaSection } from '../layouts/FormLayouts';
+import { TextAreaSection } from '../layouts/FormLayouts';
 import { transformMemberData, validateMemberData } from '../utils/common';
 import { FORM_KEYS } from '../utils/formKeys';
 import { INFO_COPY } from '../utils/content';
