+++ libs/ui/src/components/layouts/FormSegment/index.ts
@@ -1,2 +1,1 @@
-export * from './OuterLayout';
 export * from './FormSegment';
